updated TestCommand to TestMessage
